requirements.txt
bidict==0.23.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
eventlet==0.36.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-MQTT==1.2.1
Flask-SocketIO==5.3.6
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
paho-mqtt==1.6.1
ping3==4.0.8
psutil==5.9.8
python-engineio==4.9.0
python-socketio==5.11.2
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.2.1
visitor==0.1.3
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
